+++ Modules/ComputerManagementDsc/Examples/Resources/PowerShellExecutionPolicy/2-SetPowershellExecutionPolicyForMultipleScopes.ps1
@@ -22,6 +22,3 @@ Configuration Example
         } # End of ExecutionPolicyLocalMachine Resource
     } # End of Node
 } # End of Configuration
-
-#Example
-#Start-DscConfiguration Example -Wait -Verbose -Force