+++ Tests/Integration/MSFT_PowerShellExecutionPolicy.Integration.Tests.ps1
@@ -51,7 +51,7 @@ try
         #endregion
 
         It 'Should have set the resource and all the parameters should match' {
-            $current = Get-DscConfiguration | Where-Object {
+            $current = Get-DscConfiguration | Where-Object -FilterScript {
                 $_.ConfigurationName -eq "$($script:DSCResourceName)_Config"
             }
             $current.ExecutionPolicy      | Should Be 'RemoteSigned'